+++ RELEASE-NOTES.md
@@ -10,6 +10,7 @@
 * #546 Upgrade to JJava 1.0-a6
 * #547 ECharts: line chart "symbol" property
 * #548 ECharts: upgrade to 6.0.0
+* #549 ECharts: Allow custom themes (experimental)
 
 ## Release 2.0.0-M4